# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/analysis/{id}": {
            "get": {
                "description": "Retrieves contract analysis by its unique identifier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract Analysis"
                ],
                "summary": "Get analysis by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Analysis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contract analysis record",
                        "schema": {
                            "$ref": "#/definitions/repository.ContractAnalysisRecord"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/contract-analyses": {
            "get": {
                "description": "Lists contract analyses with optional filtering and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract Analysis"
                ],
                "summary": "List contract analyses",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limit for pagination",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of contract analyses",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/contract-analysis/{contract_id}": {
            "get": {
                "description": "Retrieves comprehensive analysis results for a specific contract",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract Analysis"
                ],
                "summary": "Get contract analysis by contract ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "contract_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contract analysis record",
                        "schema": {
                            "$ref": "#/definitions/repository.ContractAnalysisRecord"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/contracts/analyze-comprehensive": {
            "post": {
                "description": "Performs comprehensive contract analysis using LLM including summary extraction, payment obligations identification, percentage-based calculations, and risk assessment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract Analysis"
                ],
                "summary": "Analyze contract comprehensively",
                "parameters": [
                    {
                        "description": "Contract text to analyze",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AnalyzeContractRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comprehensive analysis result",
                        "schema": {
                            "$ref": "#/definitions/analysis.AnalyzeContractResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "analysis.AnalyzeContractResult": {
            "type": "object",
            "properties": {
                "analysis": {
                    "$ref": "#/definitions/models.ContractAnalysis"
                },
                "confidence_score": {
                    "type": "number"
                },
                "contract_id": {
                    "type": "string"
                },
                "payment_obligations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AnalysisMilestone"
                    }
                },
                "processed_at": {
                    "type": "string"
                },
                "risk_assessment": {
                    "$ref": "#/definitions/models.AnalysisRiskAssessment"
                },
                "summary": {
                    "$ref": "#/definitions/models.ContractSummary"
                },
                "validation_issues": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "handlers.AnalyzeContractRequest": {
            "type": "object",
            "required": [
                "contract_text"
            ],
            "properties": {
                "contract_text": {
                    "type": "string",
                    "example": "This is a sample contract text..."
                }
            }
        },
...
        "models.AnalysisIndividualRisk": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "party": {
                    "type": "string"
                },
                "recommendation": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.AnalysisMilestone": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "percentage": {
                    "type": "number"
                }
            }
        },
        "models.AnalysisRisk": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.AnalysisRiskAssessment": {
            "type": "object",
            "properties": {
                "compliance_score": {
                    "type": "number"
                },
                "missing_clauses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "risks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AnalysisIndividualRisk"
                    }
                },
                "suggestions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "models.ContractAnalysis": {
            "type": "object",
            "properties": {
                "buyer": {
                    "type": "string"
                },
                "buyer_address": {
                    "type": "string"
                },
                "buyer_country": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "milestones": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AnalysisMilestone"
                    }
                },
                "risk_factors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AnalysisRisk"
                    }
                },
                "seller": {
                    "type": "string"
                },
                "seller_address": {
                    "type": "string"
                },
                "seller_country": {
                    "type": "string"
                },
                "total_value": {
                    "type": "number"
                }
            }
        },
...
        "models.ContractSummary": {
            "type": "object",
            "properties": {
                "buyer_address": {
                    "type": "string"
                },
                "buyer_country": {
                    "type": "string"
                },
                "buyer_name": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "goods_nature": {
                    "type": "string"
                },
                "jurisdiction": {
                    "type": "string"
                },
                "seller_address": {
                    "type": "string"
                },
                "seller_country": {
                    "type": "string"
                },
                "seller_name": {
                    "type": "string"
                },
                "total_value": {
                    "type": "number"
                }
            }
        },
...
        "repository.ContractAnalysisRecord": {
            "type": "object",
            "properties": {
                "analysis": {
                    "$ref": "#/definitions/models.ContractAnalysis"
                },
                "confidence_score": {
                    "type": "number"
                },
                "contract_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payment_obligations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AnalysisMilestone"
                    }
                },
                "payment_obligations_json": {
                    "type": "string"
                },
                "processed_at": {
                    "type": "string"
                },
                "risk_assessment": {
                    "$ref": "#/definitions/models.AnalysisRiskAssessment"
                },
                "risk_assessment_json": {
                    "type": "string"
                },
                "summary": {
                    "$ref": "#/definitions/models.ContractSummary"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "validation_issues": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "validation_issues_json": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        }
    }
}